requirements.txt
fastapi
uvicorn
python-multipart
pydantic
langchain 
langchain_core 
langchain_community 
langgraph
langchain[groq]
docx2txt
python-dotenv
python-docx